made redis return empty array and error when no results found
Previously when accessing a model stored in redis via the API, if the model didn't exist, an empty object would be returned.  When trying to call findById the callback would give an undefined error, and an empty or null object.  This behaviour is not idiomatic nodejs and breaks the REST behaviour of returning 404 when an object doesn't exist.